+++ src/Helper/ObjectParser.php
@@ -15,6 +15,7 @@ use ReflectionProperty;
 use Yiisoft\Validator\AfterInitAttributeEventInterface;
 use Yiisoft\Validator\AttributeTranslatorInterface;
 use Yiisoft\Validator\AttributeTranslatorProviderInterface;
+use Yiisoft\Validator\Label;
 use Yiisoft\Validator\RuleInterface;
 
 use function array_key_exists;
@@ -134,6 +135,7 @@ final class ObjectParser
             'rules' => 'array',
             'reflectionAttributes' => 'array',
             'reflectionSource' => 'object',
+            'labels' => 'array',
         ],
     ])]
     private static array $cache = [];
@@ -233,6 +235,34 @@ final class ObjectParser
         $this->setCacheItem('rules', $rules);
 
         return $this->prepareRules($rules);
+    }
+
+    /**
+     * Parses labels specified via {@see Label} attributes attached to class properties.
+     *
+     * @return array<string, string>
+     */
+    public function getLabels(): array
+    {
+        if ($this->hasCacheItem('labels')) {
+            /** @var array<string, string> */
+            return $this->getCacheItem('labels');
+        }
+
+        $labels = [];
+
+        foreach ($this->getReflectionProperties() as $property) {
+            $attributes = $property->getAttributes(Label::class, ReflectionAttribute::IS_INSTANCEOF);
+            foreach ($attributes as $attribute) {
+                /** @var Label $instance */
+                $instance = $attribute->newInstance();
+                $labels[$property->getName()] = $instance->getLabel();
+            }
+        }
+
+        $this->setCacheItem('labels', $labels);
+
+        return $labels;
     }
 
     /**
@@ -415,7 +445,7 @@ final class ObjectParser
      * @return bool `true` if an item exists, `false` - if it does not or the cache is disabled in {@see $useCache}.
      */
     private function hasCacheItem(
-        #[ExpectedValues(['rules', 'reflectionProperties', 'reflectionSource'])]
+        #[ExpectedValues(['rules', 'reflectionProperties', 'reflectionSource', 'labels'])]
         string $name,
     ): bool {
         if (!$this->useCache()) {
@@ -437,7 +467,7 @@ final class ObjectParser
      * @return mixed Cache item value.
      */
     private function getCacheItem(
-        #[ExpectedValues(['rules', 'reflectionProperties', 'reflectionSource'])]
+        #[ExpectedValues(['rules', 'reflectionProperties', 'reflectionSource', 'labels'])]
         string $name,
     ): mixed {
         /** @psalm-suppress PossiblyNullArrayOffset */
@@ -451,7 +481,7 @@ final class ObjectParser
      * @param mixed $value A new value.
      */
     private function setCacheItem(
-        #[ExpectedValues(['rules', 'reflectionProperties', 'reflectionSource'])]
+        #[ExpectedValues(['rules', 'reflectionProperties', 'reflectionSource', 'labels'])]
         string $name,
         mixed $value,
     ): void {