+++ src/ValidationContext.php
@@ -58,6 +58,8 @@ final class ValidationContext
      */
     private ?string $attribute = null;
 
+    private ?string $attributeLabel = null;
+
     /**
      * @var AttributeTranslatorInterface|null Default attribute translator to use if attribute translator is not set.
      */
@@ -122,6 +124,12 @@ final class ValidationContext
     public function setAttributeTranslator(?AttributeTranslatorInterface $attributeTranslator): self
     {
         $this->attributeTranslator = $attributeTranslator;
+        return $this;
+    }
+
+    public function setAttributeLabel(?string $label): self
+    {
+        $this->attributeLabel = $label;
         return $this;
     }
 
@@ -227,6 +235,11 @@ final class ValidationContext
         return $this->attribute;
     }
 
+    public function getAttributeLabel(): ?string
+    {
+        return $this->attributeLabel;
+    }
+
     /**
      * Get translated attribute name.
      *
@@ -239,15 +252,17 @@ final class ValidationContext
             return null;
         }
 
+        $label = $this->attributeLabel ?? $this->attribute;
+
         if ($this->attributeTranslator !== null) {
-            return $this->attributeTranslator->translate($this->attribute);
+            return $this->attributeTranslator->translate($label);
         }
 
         if ($this->defaultAttributeTranslator !== null) {
-            return $this->defaultAttributeTranslator->translate($this->attribute);
+            return $this->defaultAttributeTranslator->translate($label);
         }
 
-        return $this->attribute;
+        return $label;
     }
 
     /**