+++ tests/Helper/ObjectParserTest.php
@@ -116,14 +116,18 @@ final class ObjectParserTest extends TestCase
             'b' => [new Number(min: 1)],
             'c' => [new Number(max: 2)],
         ];
+        $expectedLabels1 = ['c' => 'd'];
         $rules1 = $parser1->getRules();
+        $labels1 = $parser1->getLabels();
         $this->assertEquals($expectedRules1, $rules1);
         $cache = $cacheProperty->getValue();
         $this->assertArrayHasKey($cacheKey1, $cache);
         $this->assertArrayHasKey('rules', $cache[$cacheKey1]);
         $this->assertArrayHasKey('reflectionProperties', $cache[$cacheKey1]);
         $this->assertArrayHasKey('reflectionSource', $cache[$cacheKey1]);
+        $this->assertArrayHasKey('labels', $cache[$cacheKey1]);
         $this->assertSame($rules1, $parser1->getRules());
+        $this->assertSame($labels1,  $expectedLabels1);
 
         $parser2 = new ObjectParser(new ObjectForTestingCache2());
         $cacheKey2 = 'Yiisoft\Validator\Tests\Support\Data\ObjectForTestingCache2_7_0';
@@ -170,10 +174,14 @@ final class ObjectParserTest extends TestCase
             'b' => [new Number(min: 1)],
             'c' => [new Number(max: 2)],
         ];
+        $expectedLabels = ['c' => 'label'];
         $rules = $parser->getRules();
+        $labels = $parser->getLabels();
         $this->assertEquals($expectedRules, $rules);
+        $this->assertSame($expectedLabels, $labels);
         $this->assertArrayNotHasKey($cacheKey, $cacheProperty->getValue());
         $this->assertEquals($expectedRules, $parser->getRules());
+        $this->assertSame($expectedLabels, $parser->getLabels());
         $this->assertNotSame($rules, $parser->getRules());
         $this->assertArrayNotHasKey($cacheKey, $cacheProperty->getValue());
     }